Move Carto errors to lib, so they can be used from models
